+++ example/src/main/java/com/uploadcare/android/example/viewmodels/UploadViewModel.kt
@@ -5,13 +5,13 @@ import android.content.Context
 import android.net.Uri
 import android.view.View
 import android.view.inputmethod.InputMethodManager
+import androidx.activity.result.ActivityResultLauncher
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.Transformations
+import androidx.lifecycle.switchMap
 import com.google.android.material.textfield.TextInputEditText
 import com.uploadcare.android.example.R
-import com.uploadcare.android.example.fragments.UploadFragment
 import com.uploadcare.android.library.api.UploadcareClient
 import com.uploadcare.android.library.api.UploadcareFile
 import com.uploadcare.android.library.callbacks.UploadFileCallback
@@ -22,6 +22,7 @@ import com.uploadcare.android.widget.controller.SocialNetwork
 import com.uploadcare.android.widget.controller.UploadcareWidget
 import com.uploadcare.android.widget.controller.UploadcareWidgetResult
 import com.uploadcare.android.widget.utils.SingleLiveEvent
+import com.uploadcare.android.widget.controller.UploadcareWidgetParams
 import java.util.*
 import kotlin.math.roundToInt
 
@@ -41,13 +42,12 @@ class UploadViewModel(application: Application) : AndroidViewModel(application) 
     val status = MutableLiveData<String>()
 
     private val backgroundUploadUUID = MutableLiveData<UUID>()
-    val backgroundUploadResult : LiveData<UploadcareWidgetResult> =
-            Transformations.switchMap(backgroundUploadUUID){
-                uuid->
-                UploadcareWidget
-                        .getInstance()
-                        .backgroundUploadResult(application.applicationContext, uuid)
-    }
+    val backgroundUploadResult: LiveData<UploadcareWidgetResult> =
+        backgroundUploadUUID.switchMap { uuid ->
+            UploadcareWidget
+                .getInstance()
+                .backgroundUploadResult(application.applicationContext, uuid)
+        }
 
     val launchGetFilesCommand = SingleLiveEvent<Void>()
     val launchFilePickerCommand = SingleLiveEvent<Void>()
@@ -76,65 +76,44 @@ class UploadViewModel(application: Application) : AndroidViewModel(application) 
         }
     }
 
-    fun uploadWidgetAny(fragment: UploadFragment) {
-        val selectFileBuilder = UploadcareWidget
-                .getInstance()
-                .selectFile(fragment)
-                .cancelable(allowUploadCancelWidget.value ?: false)
-                .showProgress(showUploadProgressWidget.value ?: false)
-
-        if (backgroundUploadWidget.value == true) {
-            selectFileBuilder.backgroundUpload()
-        }
-
-        selectFileBuilder.launch()
+    fun uploadWidgetAny(uploadcareLauncher: ActivityResultLauncher<UploadcareWidgetParams>) {
+        uploadWidget(uploadcareLauncher)
     }
 
-    fun uploadWidgetInstagram(fragment: UploadFragment) {
-        val selectFileBuilder = UploadcareWidget
-                .getInstance()
-                .selectFile(fragment)
-                .style(R.style.CustomUploadCareIndigoPink)
-                .from(SocialNetwork.SOCIAL_NETWORK_INSTAGRAM)
-                .cancelable(allowUploadCancelWidget.value ?: false)
-                .showProgress(showUploadProgressWidget.value ?: false)
-
-        if (backgroundUploadWidget.value == true) {
-            selectFileBuilder.backgroundUpload()
-        }
-
-        selectFileBuilder.launch()
+    fun uploadWidgetInstagram(uploadcareLauncher: ActivityResultLauncher<UploadcareWidgetParams>) {
+        uploadWidget(
+            uploadcareLauncher = uploadcareLauncher,
+            network = SocialNetwork.SOCIAL_NETWORK_INSTAGRAM,
+            style = R.style.CustomUploadCareIndigoPink
+        )
     }
 
-    fun uploadWidgetFacebook(fragment: UploadFragment) {
-        val selectFileBuilder = UploadcareWidget
-                .getInstance()
-                .selectFile(fragment)
-                .style(R.style.CustomUploadCareGreenRed)
-                .from(SocialNetwork.SOCIAL_NETWORK_FACEBOOK)
-                .cancelable(allowUploadCancelWidget.value ?: false)
-                .showProgress(showUploadProgressWidget.value ?: false)
-
-        if (backgroundUploadWidget.value == true) {
-            selectFileBuilder.backgroundUpload()
-        }
-
-        selectFileBuilder.launch()
+    fun uploadWidgetFacebook(uploadcareLauncher: ActivityResultLauncher<UploadcareWidgetParams>) {
+        uploadWidget(
+            uploadcareLauncher = uploadcareLauncher,
+            network = SocialNetwork.SOCIAL_NETWORK_FACEBOOK,
+            style = R.style.CustomUploadCareGreenRed
+        )
     }
 
-    fun uploadWidgetDropbox(fragment: UploadFragment) {
-        val selectFileBuilder = UploadcareWidget
-                .getInstance()
-                .selectFile(fragment)
-                .from(SocialNetwork.SOCIAL_NETWORK_DROPBOX)
-                .cancelable(allowUploadCancelWidget.value ?: false)
-                .showProgress(showUploadProgressWidget.value ?: false)
+    fun uploadWidgetDropbox(uploadcareLauncher: ActivityResultLauncher<UploadcareWidgetParams>) {
+        uploadWidget(uploadcareLauncher, SocialNetwork.SOCIAL_NETWORK_DROPBOX)
+    }
 
-        if (backgroundUploadWidget.value == true) {
-            selectFileBuilder.backgroundUpload()
-        }
+    private fun uploadWidget(
+        uploadcareLauncher: ActivityResultLauncher<UploadcareWidgetParams>,
+        network: SocialNetwork? = null,
+        style: Int = -1,
+    ) {
+        val params = UploadcareWidgetParams(
+            style = style,
+            network = network,
+            cancelable = allowUploadCancelWidget.value ?: false,
+            showProgress = showUploadProgressWidget.value ?: false,
+            backgroundUpload = backgroundUploadWidget.value ?: false
+        )
 
-        selectFileBuilder.launch()
+        uploadcareLauncher.launch(params)
     }
 
     fun onUploadResult(result: UploadcareWidgetResult) {
@@ -142,47 +121,18 @@ class UploadViewModel(application: Application) : AndroidViewModel(application) 
             if (result.uploadcareFile != null) {
                 showProgressOrResult(false, result.uploadcareFile.toString())
             } else if (result.backgroundUploadUUID != null) {
-                showProgressOrResult(false,
-                        "Uploading in background: ${result.backgroundUploadUUID.toString()}")
+                showProgressOrResult(
+                    false,
+                    "Uploading in background: ${result.backgroundUploadUUID.toString()}"
+                )
 
-                if(result.backgroundUploadUUID != backgroundUploadUUID.value){
-                    backgroundUploadUUID.value = result.backgroundUploadUUID
+                if (result.backgroundUploadUUID != backgroundUploadUUID.value) {
+                    backgroundUploadUUID.value = result.backgroundUploadUUID!!
                 }
             }
         } else {
             showProgressOrResult(false, result.exception?.message.toString())
         }
-    }
-
-    /**
-     * Uploads file from {@link Uri} using UploadcareClient.
-     *
-     * @param fileUri file {@link Uri}
-     */
-    fun uploadFile(fileUri: Uri) {
-        showProgressOrResult(true, getContext().getString(R.string.activity_main_status_uploading))
-        allowUploadPause.value = true
-        uploader = FileUploader(client, fileUri, getContext()).store(true)
-        uploader!!.uploadAsync(object : UploadFileCallback {
-            override fun onFailure(e: UploadcareApiException) {
-                allowUploadPause.value = false
-                showProgressOrResult(false, e.message ?: "")
-            }
-
-            override fun onProgressUpdate(
-                    bytesWritten: Long,
-                    contentLength: Long,
-                    progress: Double) {
-                if (showUploadProgress.value == true) {
-                    uploadProgress.value = (progress * 100).roundToInt()
-                }
-            }
-
-            override fun onSuccess(result: UploadcareFile) {
-                allowUploadPause.value = false
-                showProgressOrResult(false, result.toString())
-            }
-        })
     }
 
     /**
@@ -193,37 +143,40 @@ class UploadViewModel(application: Application) : AndroidViewModel(application) 
     fun uploadFiles(filesUriList: List<Uri>) {
         showProgressOrResult(true, getContext().getString(R.string.activity_main_status_uploading))
         allowUploadPause.value = true
-        multipleUploader = MultipleFilesUploader(client, filesUriList, getContext()).store(true)
-        multipleUploader!!.uploadAsync(object : UploadFilesCallback {
+        multipleUploader = MultipleFilesUploader(client, filesUriList, getContext())
+            .store(true)
+            .also { multipleFilesUploader ->
+                multipleFilesUploader.uploadAsync(object : UploadFilesCallback {
+                    override fun onFailure(e: UploadcareApiException) {
+                        allowUploadPause.value = false
+                        showProgressOrResult(false, e.message ?: "")
+                    }
 
-            override fun onFailure(e: UploadcareApiException) {
-                allowUploadPause.value = false
-                showProgressOrResult(false, e.message ?: "")
-            }
+                    override fun onProgressUpdate(
+                        bytesWritten: Long,
+                        contentLength: Long,
+                        progress: Double
+                    ) {
+                        if (showUploadProgress.value == true) {
+                            uploadProgress.value = (progress * 100).roundToInt()
+                        }
+                    }
 
-            override fun onProgressUpdate(
-                    bytesWritten: Long,
-                    contentLength: Long,
-                    progress: Double) {
-                if (showUploadProgress.value == true) {
-                    uploadProgress.value = (progress * 100).roundToInt()
-                }
+                    override fun onSuccess(result: List<UploadcareFile>) {
+                        allowUploadPause.value = false
+                        val resultStringBuilder = StringBuilder()
+                        for (file in result) {
+                            resultStringBuilder.append(file.toString())
+                                .append(System.getProperty("line.separator"))
+                                .append(System.getProperty("line.separator"))
+                        }
+                        showProgressOrResult(false, resultStringBuilder.toString())
+                    }
+                })
             }
-
-            override fun onSuccess(result: List<UploadcareFile>) {
-                allowUploadPause.value = false
-                val resultStringBuilder = StringBuilder()
-                for (file in result) {
-                    resultStringBuilder.append(file.toString())
-                            .append(System.getProperty("line.separator"))
-                            .append(System.getProperty("line.separator"))
-                }
-                showProgressOrResult(false, resultStringBuilder.toString())
-            }
-        })
     }
 
-    fun cancelUpload(){
+    fun cancelUpload() {
         uploader?.cancel()
         multipleUploader?.cancel()
         uploader = null
@@ -243,8 +196,8 @@ class UploadViewModel(application: Application) : AndroidViewModel(application) 
             uploadPaused.value = uploader.isPaused
         }
 
-        val multipleFilesUploader: MultipleFilesUploader?
-                = multipleUploader as? MultipleFilesUploader
+        val multipleFilesUploader: MultipleFilesUploader? =
+            multipleUploader as? MultipleFilesUploader
         multipleFilesUploader?.let {
             if (multipleFilesUploader.isPaused) {
                 multipleFilesUploader.resume()
@@ -263,25 +216,29 @@ class UploadViewModel(application: Application) : AndroidViewModel(application) 
      */
     private fun uploadFromUrl(client: UploadcareClient, sourceUrl: String) {
         showProgressOrResult(true, getContext().getString(R.string.activity_main_status_uploading))
-        uploader = UrlUploader(client, sourceUrl).store(true)
-        uploader!!.uploadAsync(object : UploadFileCallback {
-            override fun onFailure(e: UploadcareApiException) {
-                showProgressOrResult(false, e.message ?: "")
-            }
+        uploader = UrlUploader(client, sourceUrl)
+            .store(true)
+            .also { urlUploader ->
+                urlUploader.uploadAsync(object : UploadFileCallback {
+                    override fun onFailure(e: UploadcareApiException) {
+                        showProgressOrResult(false, e.message ?: "")
+                    }
 
-            override fun onProgressUpdate(
-                    bytesWritten: Long,
-                    contentLength: Long,
-                    progress: Double) {
-                if (showUploadProgress.value == true) {
-                    uploadProgress.value = (progress * 100).roundToInt()
-                }
-            }
+                    override fun onProgressUpdate(
+                        bytesWritten: Long,
+                        contentLength: Long,
+                        progress: Double
+                    ) {
+                        if (showUploadProgress.value == true) {
+                            uploadProgress.value = (progress * 100).roundToInt()
+                        }
+                    }
 
-            override fun onSuccess(result: UploadcareFile) {
-                showProgressOrResult(false, result.toString())
+                    override fun onSuccess(result: UploadcareFile) {
+                        showProgressOrResult(false, result.toString())
+                    }
+                })
             }
-        })
     }
 
     /**
@@ -291,7 +248,7 @@ class UploadViewModel(application: Application) : AndroidViewModel(application) 
      */
     private fun checkUrl(url: String?): Boolean {
         return if (url != null && url.isNotEmpty()) {
-            urlError.value = null
+            urlError.value = ""
             true
         } else {
             urlError.value = getContext().getString(R.string.activity_main_hint_upload_url)