+++ library/src/main/java/com/uploadcare/android/library/api/UploadcareFile.kt
@@ -3,7 +3,7 @@ package com.uploadcare.android.library.api
 import android.os.Parcelable
 import com.squareup.moshi.Json
 import com.uploadcare.android.library.urls.CdnPathBuilder
-import kotlinx.android.parcel.Parcelize
+import kotlinx.parcelize.Parcelize
 import java.net.URI
 import java.util.*