+++ widget/src/main/java/com/uploadcare/android/widget/controller/UploadcareWidget.kt
@@ -17,8 +17,9 @@ import com.uploadcare.android.widget.R
 import com.uploadcare.android.widget.activity.UploadcareActivity
 import com.uploadcare.android.widget.interfaces.SocialApi
 import com.uploadcare.android.widget.utils.SingletonHolder
+import com.uploadcare.android.widget.utils.getSupportParcelable
 import com.uploadcare.android.widget.worker.FileUploadWorker
-import kotlinx.android.parcel.Parcelize
+import kotlinx.parcelize.Parcelize
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import java.util.*
@@ -59,6 +60,10 @@ class UploadcareWidget private constructor(context: Context) {
      * @return Builder where you can specify all required parameters and launch selection
      * and upload.
      */
+    @Deprecated(
+        "Use UploadcareActivityResultContract with registerForActivityResult(ActivityResultContract, ActivityResultCallback) from Activity or Fragment",
+        ReplaceWith("registerForActivityResult(UploadcareActivityResultContract) {}", "com.uploadcare.android.widget.controller.UploadcareActivityResultContract")
+    )
     fun selectFile(activity: Activity): Builder {
         return Builder(activity)
     }
@@ -71,6 +76,10 @@ class UploadcareWidget private constructor(context: Context) {
      * @return Builder where you can specify all required parameters and launch selection
      * and upload.
      */
+    @Deprecated(
+        "Use UploadcareActivityResultContract with registerForActivityResult(ActivityResultContract, ActivityResultCallback) from Activity or Fragment",
+        ReplaceWith("registerForActivityResult(UploadcareActivityResultContract) {}", "com.uploadcare.android.widget.controller.UploadcareActivityResultContract")
+    )
     fun selectFile(fragment: Fragment): Builder {
         return Builder(fragment)
     }
@@ -137,6 +146,10 @@ class UploadcareWidget private constructor(context: Context) {
         return result
     }
 
+    @Deprecated(
+        "Use UploadcareActivityResultContract with registerForActivityResult(ActivityResultContract, ActivityResultCallback) from Activity or Fragment",
+        ReplaceWith("registerForActivityResult(UploadcareActivityResultContract) {}", "com.uploadcare.android.widget.controller.UploadcareActivityResultContract")
+    )
     class Builder private constructor(private val fragment: Fragment? = null,
                                       private val activity: Activity? = null) {
 
@@ -243,6 +256,10 @@ class UploadcareWidget private constructor(context: Context) {
          * To get result with info about uploaded file or error, override onActivityResult() method,
          * and use code below to get result
          * val result = UploadcareWidgetResult.fromIntent(data)
+         *
+         * Notice: Since [androidx.fragment.app.Fragment.startActivityForResult] was deprecated, you
+         * should use [androidx.fragment.app.Fragment.registerForActivityResult] with
+         * [com.uploadcare.android.widget.controller.UploadcareActivityResultContract] instead.
          */
         fun launch() {
             if (fragment != null) {
@@ -330,9 +347,11 @@ data class UploadcareWidgetResult(val uploadcareFile: UploadcareFile? = null,
     fun isFailed() = exception != null
 
     companion object {
+
+        private const val KEY_RESULT = "result"
+
         @JvmStatic
-        fun fromIntent(intent: Intent?): UploadcareWidgetResult? {
-            return intent?.getParcelableExtra("result")
-        }
+        fun fromIntent(intent: Intent?): UploadcareWidgetResult? =
+            intent?.extras?.getSupportParcelable(KEY_RESULT, UploadcareWidgetResult::class.java)
     }
 }