+++ widget/src/main/java/com/uploadcare/android/widget/data/SocialSource.kt
@@ -2,17 +2,16 @@ package com.uploadcare.android.widget.data
 
 import android.content.Context
 import android.os.Parcelable
-import android.preference.PreferenceManager
 import android.webkit.CookieManager
+import androidx.preference.PreferenceManager
 import com.squareup.moshi.Json
 import com.uploadcare.android.widget.R
-import kotlinx.android.parcel.Parcelize
-import kotlinx.android.parcel.RawValue
+import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class SocialSource(@Json(name = "root_chunks") val rootChunks: List<Chunk>,
                         val name: String,
-                        val urls: @RawValue Urls) : Parcelable {
+                        val urls: Urls) : Parcelable {
 
     fun saveCookie(context: Context, cookie: String) {
         PreferenceManager.getDefaultSharedPreferences(context)