+++ widget/src/main/java/com/uploadcare/android/widget/data/SocialSourcesResponse.kt
@@ -1,7 +1,7 @@
 package com.uploadcare.android.widget.data
 
 import android.os.Parcelable
-import kotlinx.android.parcel.Parcelize
+import kotlinx.parcelize.Parcelize
 
 @Parcelize
 data class SocialSourcesResponse(val sources: List<SocialSource>) : Parcelable