+++ widget/src/main/java/com/uploadcare/android/widget/fragment/UploadcareFilesFragment.kt
@@ -6,6 +6,7 @@ import android.os.Bundle
 import android.view.*
 import android.widget.AdapterView
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.MenuProvider
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.get
@@ -23,17 +24,12 @@ import com.uploadcare.android.widget.utils.NavigationHelper
 import com.uploadcare.android.widget.viewmodels.UploadcareFilesViewModel
 
 class UploadcareFilesFragment : Fragment(), AdapterView.OnItemSelectedListener,
-        OnFileActionsListener, OnAuthListener, CancelUploadListener {
+        OnFileActionsListener, OnAuthListener, CancelUploadListener, MenuProvider {
 
     private lateinit var binding: UcwFragmentFilesBinding
     private lateinit var viewModel: UploadcareFilesViewModel
 
     private val args: UploadcareFilesFragmentArgs by navArgs()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setHasOptionsMenu(true)
-    }
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View {
@@ -115,13 +111,19 @@ class UploadcareFilesFragment : Fragment(), AdapterView.OnItemSelectedListener,
         return binding.root
     }
 
-    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
-        super.onCreateOptionsMenu(menu, inflater)
-        inflater.inflate(R.menu.ucw_social_actions, menu)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        requireActivity().addMenuProvider(
+            this,
+            viewLifecycleOwner
+        )
     }
 
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        return when (item.itemId) {
+    override fun onCreateMenu(menu: Menu, menuInflater: MenuInflater) {
+        menuInflater.inflate(R.menu.ucw_social_actions, menu)
+    }
+
+    override fun onMenuItemSelected(menuItem: MenuItem): Boolean =
+        when (menuItem.itemId) {
             android.R.id.home -> {
                 checkBackStack()
                 true
@@ -130,9 +132,8 @@ class UploadcareFilesFragment : Fragment(), AdapterView.OnItemSelectedListener,
                 viewModel.signOut()
                 true
             }
-            else -> super.onOptionsItemSelected(item)
+            else -> false
         }
-    }
 
     override fun onNothingSelected(parent: AdapterView<*>?) {
         // Ignore.